requirements.txt
ajpy==0.0.4
apparmor==2.13.2
asn1crypto==0.24.0
astroid==2.2.5
backdoor-factory==0.0.0
bcrypt==3.1.6
beautifulsoup4==4.7.1
binwalk==2.1.2
bleach==3.1.0
blinker==1.4
Brotli==1.0.7
cachetools==3.1.1
capstone==3.0.5
certifi==2018.8.24
chardet==3.0.4
Click==7.0
colorama==0.3.7
crcelk==1.3
cryptography==2.6.1
cupshelpers==1.0
cycler==0.10.0
Cython==0.29.2
decorator==4.3.0
defusedxml==0.5.0
distro==1.3.0
Django==2.2.2
django-filter==2.1.0
djangorestframework==3.9.4
dnspython==1.16.0
ecdsa==0.13
Electrum==3.2.3
entrypoints==0.3
Flask==1.0.2
future==0.16.0
google-api-python-client==1.7.9
google-auth==1.6.3
google-auth-httplib2==0.0.3
gpg==1.12.0
h11==0.8.1
h2==3.0.1
hashID==3.1.4
hpack==3.0.0
html5lib==1.0.1
httplib2==0.11.3
hyperframe==5.1.0
idna==2.6
iotop==0.6
IPy==0.83
ipykernel==4.9.0
ipython==5.8.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
isort==4.3.20
itsdangerous==0.24
Jinja2==2.10
jsonrpclib-pelix==0.3.1
jsonschema==2.6.0
jupyter-client==5.2.3
jupyter-core==4.4.0
kaitaistruct==0.8
keyring==17.1.1
keyrings.alt==3.1.1
kiwisolver==1.0.1
lazy-object-proxy==1.4.1
ldap3==2.5.1
LibAppArmor==2.13.2
lxml==4.3.2
Mako==1.0.7
Markdown==3.1.1
MarkupSafe==1.1.0
mate-tweak==18.10.2
matplotlib==3.0.2
mccabe==0.6.1
meld==3.20.0
mistune==0.8.4
mitmproxy==4.0.4
mutagen==1.40.0
mysqlclient==1.3.10
nassl==2.1.2
nbconvert==5.4.0
nbformat==4.4.0
notebook==5.7.4
numpy==1.16.2
nyx==2.1.0
olefile==0.46
onboard==1.4.1
onioncircuits==0.5
onionshare==1.3.2
opencv-python==4.1.0.25
pandocfilters==1.4.2
paramiko==2.4.2
passlib==1.7.1
patator==0.7
pefile==2018.8.8
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.4.1
pipenv==2018.11.26
plotly==3.6.1
pluginbase==1.0.0
ply==3.11
prometheus-client==0.6.0
prompt-toolkit==1.0.15
protobuf==3.6.1
psutil==5.5.1
psycopg2==2.7.7
pyaes==1.6.1
pyasn1==0.4.2
pyasn1-modules==0.2.5
pycairo==1.16.2
pycountry==17.5.14
pycrypto==2.6.1
pycryptodomex==3.6.1
pycups==1.9.73
pycurl==7.43.0.2
Pygments==2.3.1
PyGObject==3.30.4
pyinotify==0.9.6
pylint==2.3.1
PyNaCl==1.3.0
PyOpenGL==3.1.0
pyOpenSSL==19.0.0
pyparsing==2.2.0
pyperclip==1.6.4
pyqtgraph==0.10.0
pyserial==3.4
PySimpleSOAP==1.16.2
pysmbc==1.0.15.6
pysmi==0.3.2
pysnmp==4.4.6
PySocks==1.6.8
python-apt==1.8.4
python-dateutil==2.7.3
python-debian==0.1.34
python-debianbts==2.8.2
python-xlib==0.23
pytz==2019.1
pyxattr==0.6.1
pyxdg==0.25
pyzmq==17.1.2
qrcode==6.1
QtPy==1.3.1
reportbug==7.5.2
requests==2.21.0
retrying==1.3.3
rsa==4.0
ruamel.yaml==0.15.34
scipy==1.1.0
SecretStorage==2.3.1
selenium==3.141.0
Send2Trash==1.5.0
setproctitle==1.1.10
simplegeneric==0.8.1
simplejson==3.16.0
six==1.12.0
smoke-zephyr==1.4.1
sortedcontainers==2.0.4
soupsieve==1.8
sqlparse==0.3.0
sslyze==2.0.6
stem==1.7.1
tabulate==0.8.2
termcolor==1.1.0
terminado==0.8.1
termineter==1.0.4
testpath==0.4.2
tld==0.9.1
tls-parser==1.2.1
torbrowser-launcher==0.3.1
tornado==5.1.1
traitlets==4.3.2
typed-ast==1.4.0
uritemplate==3.0.0
urllib3==1.24.1
urwid==2.0.1
virtualenv==16.6.1
virtualenv-clone==0.5.3
wafw00f==0.9.5
wapiti3==3.0.1
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
wfuzz==2.3.4
wrapt==1.11.1
wsproto==0.11.0
yaswfp==0.9.3
youtube-dl==2019.1.17
